+++ tests/testthat/test-aux--gis.R
@@ -1,3 +1,23 @@
-test_that("plot_albufera_clusters(): no error with default args", {
+test_that("info_clusters() succeeds", {
+  expect_no_error(info_clusters())
+})
+
+test_that("info_clusters(include_geometry = TRUE) succeeds", {
+  expect_no_error(info_clusters(include_geometry = TRUE))
+})
+
+test_that("info_ditches() succeeds", {
+  expect_no_error(info_ditches())
+})
+
+test_that("info_ditches(include_geometry = TRUE) succeeds", {
+  expect_no_error(info_ditches(include_geometry = TRUE))
+})
+
+
+
+
+test_that("plot_albufera_clusters(): no error or warning with default args", {
   expect_no_error(plot_albufera_clusters())
+  expect_no_warning(plot_albufera_clusters())
 })